requirements.txt
keras==2.10.0
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
streamlit==1.13.0
tensorflow==2.10.0
